Initial source files of RHO2_OPS
A Makefile is provided to compile the code. It can be easily adapted for your preferred compiler (OpenMP is required for its compilation).  

chimpanC is an interface to transform 2-RDM matrices from the molecular orbital basis to the primitive basis. This code is compatible with DMN code of Dr. Matito and with DoNOF code of Dr. Piris. Although, 2-RDM and the .basis files can be generated using your own interface. 

RHO2_OPS requires a .basis (with the information about the primitives basis set) and a 2-RDM (in primitives as well). To compute: intracules, extracules and total position spread. The first two quantities are computed using the algorithm of Cioslowksi and Liu.